user endpoint; require auth; cleanup
added a /kratos/user endpoint to return the logged in user

added a couch_utils.rewrite method to call a rewrite. cleaning up
a couple of endpoints.

added a _users/lib/helpers with a sanitize_user method to remove
sensitive info from an arbitrary user doc. used in all lists/shows.

now couch middleware returns a 401 unauthorized if you are not
logged in.