make install work with hardened image
install epel

`./configure` and `make` with couchdb user; only `make install`
with root. that way most permissions are fixed.

make all dirs and subdirs in /usr/local a+x -
should just be the ones we actually use.

make all dirs and subdirs in proj dir a+rx so static files can be
served