removed unused models; cleaned migrations; fixed db permissions
we cannot chown files in a shared vagrant folder; so instead we use
a variable to specify the web user. On vagrant, we make that user
Vagrant. We chown the db to be owned by the web user.

removed all the kratos-specific models; moved the user model from
kratos to core, and removed kratos-specific methods. removed
the osw user extension because we no longer needed it after
switching entire signup process to pipelines.

removed all the old migrations and replaced with a single initial.

There is still a fair bit of kratos cruft, and a whole bunch of
unused views. but it was most important to get the models cleaned
up before testing.