Adding create web user task;
The ordering of the django and supervisord roles is creating
a problem where the web_user is not created on the server
when the django role is run.  The user is created by supervisord.
Rather than creating a separate role specifically to create this
one system user, I simply copied it into the role for django,
since it's more fair to say the web_user is a dependency for both
roles.